* Improved __ewl_button_update_label a bit.  * Made the size of the Ewl_Text etox area allways be the actual size    of it without breaking rows.  * Added a button test part to the test program.  * Fixed some minor bugs in the Ewl_CheckButton & Ewl_RadioButton    widget.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@5241 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/ewl/src/ewl_text.h
@@ -31,10 +31,10 @@ void ewl_text_get_color(Ewl_Widget * widget, int *r, int *g, int *b, int *a);
 void ewl_text_set_style(Ewl_Widget * w, char * s);
 void ewl_text_get_text_geometry(Ewl_Widget * widget, double *x, double *y,
 				double *w, double *h);
-void ewl_text_get_letter_geometry(Ewl_Widget * widget, int index, int *x,
-				  int *y, int *w, int *h);
-void ewl_text_get_letter_geometry_at(Ewl_Widget * widget, int x, int y,
-				     int *tx, int *ty, int *tw, int *th);
+void ewl_text_get_letter_geometry(Ewl_Widget * widget, int index, double *x,
+				  double *y, double *w, double *h);
+void ewl_text_get_letter_geometry_at(Ewl_Widget * widget, double x, double y,
+				     double *tx, double *ty, double *tw, double *th);
 void ewl_text_set_alignment(Ewl_Widget * w, Ewl_Alignment a);
 
 #endif
